archive/python-gpu/queryserver/requirements.txt
haystack-ai
aiohttp
numpy
tqdm
sentence-transformers
accelerate